Add - Introduce resource module to seperate common resources in one place

+++ domain/src/main/java/com/ajijul/elnaz/domain/usecases/category/GetAllSubCategoriesUseCase.kt
@@ -6,6 +6,6 @@ import kotlinx.coroutines.flow.Flow
 
 class GetAllSubCategoriesUseCase(val categoryRepository: CategoryRepository) {
 
-    operator fun invoke(parentCategoryId: Int): Flow<List<CategoryModel>> = categoryRepository.getAllSubCategories(parentCategoryId)
+    operator fun invoke(parentCategoryId: Long): Flow<List<CategoryModel>> = categoryRepository.getAllSubCategories(parentCategoryId)
 
 }
